renamed migrations folder to seeds
