+++ db/index.js
@@ -1,12 +1,7 @@
 const
 elasticsearch = require('elasticsearch'),
 init          = require('./init'),
-models        = require('./models/index'),
 client        = new elasticsearch.Client({host: 'localhost:9200', log: 'trace'});
-
-module.exports = {
-  models
-}
 
 init.testConnect(client);
 // create index if not exist
