+++ models/response.js
@@ -1,0 +1,11 @@
+/*
+  type: "info",
+  id: <program_id> ex: "AdultHealthBenefit",
+  text: HTML fragment that will be rendered by VSAQ
+*/
+
+exports.mapping = {
+  type: {type: "string"},
+  id: {type:"string"},
+  text: {type:"string"}
+}
